requirements.txt
azure-devops==7.1.0b4
msrest==0.7.1
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
playwright>=1.40.0